discord_app/requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.1.0
discord.py==1.7.3
django-environ==0.9.0
feedparser==6.0.10
idna==2.10
multidict==4.7.6
requests==2.28.1
sgmllib3k==1.0.0
urllib3==1.25.11
websockets==8.1
yarl==1.5.1